Updated compute_pos() function
I am a little unsure how to characterize the covariance of the state estimate (I can't quite remember if there is a covariance matrix associated with the state estimate). If there is no covariance associated with the state estimate, the covariance of the position of Agent n is just the covariance of the range measurement.